Fixed crops chart permissions by moving to another controller
